requirements.txt
RPi.GPIO
flask
flask_cors
pyserial
